+++ demo/client/index.js
@@ -34,8 +34,9 @@ import {
   withChildren,
   withChild,
   EMPTY_OBJECT,
+  logProps,
+  withElement,
 } from '../../src'
-
 import { Resources } from './resources'
 
 const Text = compose(
@@ -290,6 +291,7 @@ export const Toggle = compose(
   ),
   renameProp('onPush', 'onChange'),
   cyclable,
+  logProps(),
 )(function Toggle({ value, onCycle }) {
   return $(
     'div',
@@ -303,6 +305,17 @@ export const Toggle = compose(
     $('p', null, value ? 'ON' : 'OFF'),
   )
 })
+
+const Article = withElement({ header: 'h1', body: 'p' }, (props, name) => ({
+  children: props.value[name],
+}))(({ children = EMPTY_OBJECT }) =>
+  $(
+    'div',
+    null,
+    $('div', null, children.header),
+    $('div', null, children.body),
+  ),
+)
 
 export const App = compose(
   withProps({
@@ -340,6 +353,7 @@ export const App = compose(
     $(Color, property('color')),
     $('h2', null, 'Delayed'),
     $(Toggle, { ...property('toggle'), delay: 2000 }),
+    $(Article, { value: { header: 'Title', body: 'Content' } }),
   )
 })